Move save_gradient_sampling to scilpy.io
