+++ scilpy/io/save_gradient_sampling.py
